dashboard/requirements.txt
flask==2.2.3
flask-socketio==5.3.3
kafka-python==2.0.2
python-socketio==5.8.0
eventlet==0.33.3
